feat: fallback dirigeant on recherche entreprise
