+++ app/(header-default)/dirigeants/[slug]/_component/sections/is-personne-morale.ts
@@ -1,4 +1,4 @@
-import { IEtatCivil, IPersonneMorale } from '#models/immatriculation';
+import { IEtatCivil, IPersonneMorale } from '#models/rne/types';
 
 /**
  * Weird bug happennig here. Webpack build fail when this function is in model/dirigeants.ts
