+++ app/(header-default)/elus/_component/section/elus-section.tsx
@@ -1,11 +1,11 @@
-import React from 'react';
 import NonRenseigne from '#components/non-renseigne';
 import { Section } from '#components/section';
 import { FullTable } from '#components/table/full';
 import { EAdministration } from '#models/administrations/EAdministration';
 import { isCollectiviteTerritoriale, IUniteLegale } from '#models/core/types';
-import { IEtatCivil } from '#models/immatriculation';
+import { IEtatCivil } from '#models/rne/types';
 import { capitalize, formatDatePartial } from '#utils/helpers';
+import React from 'react';
 
 /**
  * Elus section
