+++ clients/recherche-entreprise/index.ts
@@ -5,23 +5,12 @@ import constants from '#models/constants';
 import { createEtablissementsList } from '#models/core/etablissements-list';
 import { IETATADMINSTRATIF, estActif } from '#models/core/etat-administratif';
 import {
-  IEtablissement,
   NotEnoughParamsException,
-  createDefaultEtablissement,
   createDefaultUniteLegale,
 } from '#models/core/types';
-import { IEtatCivil, IPersonneMorale } from '#models/immatriculation';
 import { ISearchResults } from '#models/search';
 import SearchFilterParams from '#models/search-filter-params';
-import {
-  Siret,
-  extractNicFromSiret,
-  extractSirenFromSiret,
-  formatFirstNames,
-  parseIntWithDefaultValue,
-  verifySiren,
-  verifySiret,
-} from '#utils/helpers';
+import { parseIntWithDefaultValue, verifySiren } from '#utils/helpers';
 import {
   libelleFromCategoriesJuridiques,
   libelleFromCodeNAFWithoutNomenclature,
@@ -32,13 +21,14 @@ import {
   statuDiffusionFromStatutDiffusionInsee,
 } from '#utils/helpers/insee-variables';
 import { httpGet } from '#utils/network';
+import { IResult, ISearchResponse } from './interface';
 import {
-  IDirigeant,
-  IMatchingEtablissement,
-  IResult,
-  ISearchResponse,
-  ISiege,
-} from './interface';
+  mapToDirigeantModel,
+  mapToElusModel,
+  mapToEtablissement,
+  mapToImmatriculation,
+  mapToSiege,
+} from './mapToDomain';
 
 type ClientSearchRechercheEntreprise = {
   searchTerms: string;
@@ -175,18 +165,6 @@ const mapToUniteLegale = (result: IResult, pageEtablissements: number) => {
     type_siae = '',
   } = complements || {};
 
-  const {
-    date_debut_activite = '',
-    date_immatriculation = '',
-    date_radiation = '',
-    duree_personne_morale,
-    nature_entreprise = [],
-    date_cloture_exercice = '',
-    capital_social,
-    capital_variable = false,
-    devise_capital = '',
-  } = immatriculation || {};
-
   const nomComplet = (result.nom_complet || 'Nom inconnu').toUpperCase();
 
   const siren = verifySiren(result.siren);
@@ -273,18 +251,7 @@ const mapToUniteLegale = (result: IResult, pageEtablissements: number) => {
       estEntrepriseInclusive: est_siae,
       typeEntrepriseInclusive: type_siae,
     },
-    immatriculation: {
-      dateDebutActivite: date_debut_activite ?? '',
-      dateImmatriculation: date_immatriculation ?? '',
-      dateRadiation: date_radiation ?? '',
-      duree: duree_personne_morale ?? 0,
-      natureEntreprise: nature_entreprise || [],
-      dateCloture: date_cloture_exercice ?? '',
-      isPersonneMorale: !!capital_social,
-      capital: capital_social ?? 0,
-      estCapitalVariable: capital_variable ?? false,
-      deviseCapital: devise_capital ?? '',
-    },
+    immatriculation: mapToImmatriculation(immatriculation),
     association: {
       idAssociation: identifiant_association,
       data: null,
@@ -296,150 +263,6 @@ const mapToUniteLegale = (result: IResult, pageEtablissements: number) => {
     dateMiseAJourInpi: date_mise_a_jour_rne || '',
     dateFermeture: date_fermeture ?? '',
     listeIdcc: liste_idcc || [],
-  };
-};
-
-const mapToSiege = (
-  siege: IResult['siege'],
-  est_entrepreneur_individuel: boolean
-) => {
-  if (!siege || Object.keys(siege).length === 0 || !siege.siret) {
-    return {
-      ...createDefaultEtablissement(),
-      siret: '' as Siret,
-    };
-  }
-  return mapToEtablissement(siege, est_entrepreneur_individuel);
-};
-
-const mapToDirigeantModel = (
-  dirigeant: IDirigeant
-): IEtatCivil | IPersonneMorale => {
-  const {
-    siren = '',
-    sigle = '',
-    denomination = '',
-    prenoms = '',
-    nom = '',
-    qualite = '',
-  } = dirigeant;
-
-  if (!!siren) {
-    return {
-      siren,
-      denomination: `${denomination}${sigle ? ` (${sigle})` : ''}`,
-      role: qualite,
-    } as IPersonneMorale;
-  }
-
-  return {
-    sexe: null,
-    nom: (nom || '').toUpperCase(),
-    prenom: formatFirstNames((prenoms || '').split(' '), 1),
-    role: qualite,
-    dateNaissancePartial: '',
-    lieuNaissance: '',
-  };
-};
-
-const mapToElusModel = (eluRaw: any): IEtatCivil => {
-  const { nom, prenoms, annee_de_naissance, fonction, sexe } = eluRaw;
-
-  return {
-    sexe,
-    nom: (nom || '').toUpperCase(),
-    prenom: formatFirstNames((prenoms || '').split(' '), 1),
-    role: fonction,
-    dateNaissancePartial: annee_de_naissance,
-    lieuNaissance: '',
-  };
-};
-
-const mapToEtablissement = (
-  etablissement: ISiege | IMatchingEtablissement,
-  estEntrepreneurIndividuel: boolean
-): IEtablissement => {
-  const {
-    siret,
-    latitude = '0',
-    longitude = '0',
-    code_postal = '',
-    libelle_commune = '',
-    adresse,
-    liste_enseignes,
-    etat_administratif,
-    est_siege = false,
-    ancien_siege = false,
-    nom_commercial = '',
-    activite_principale = '',
-    date_creation = '',
-    date_debut_activite = '',
-    date_fermeture = '',
-    tranche_effectif_salarie = '',
-    caractere_employeur = '',
-    annee_tranche_effectif_salarie = '',
-    liste_finess = [],
-    liste_id_bio = [],
-    liste_idcc = [],
-    liste_id_organisme_formation = [],
-    liste_rge = [],
-    liste_uai = [],
-    statut_diffusion_etablissement,
-  } = etablissement;
-
-  const enseigne = (liste_enseignes || []).join(' ');
-
-  const adressePostale = adresse
-    ? `${
-        enseigne ? `${enseigne}, ` : nom_commercial ? `${nom_commercial}, ` : ''
-      }${adresse}`
-    : '';
-
-  const etatAdministratif = etatFromEtatAdministratifInsee(
-    etat_administratif,
-    siret
-  );
-
-  return {
-    ...createDefaultEtablissement(),
-    siren: extractSirenFromSiret(siret),
-    enseigne,
-    nic: extractNicFromSiret(siret),
-    siret: verifySiret(siret),
-    adresse,
-    codePostal: code_postal,
-    commune: libelle_commune,
-    adressePostale,
-    trancheEffectif:
-      caractere_employeur === 'N'
-        ? caractere_employeur
-        : tranche_effectif_salarie,
-    anneeTrancheEffectif: annee_tranche_effectif_salarie,
-    latitude,
-    longitude,
-    estSiege: est_siege,
-    ancienSiege: ancien_siege,
-    etatAdministratif,
-    statutDiffusion: statuDiffusionFromStatutDiffusionInsee(
-      statut_diffusion_etablissement || 'O',
-      siret
-    ),
-    denomination: nom_commercial,
-    libelleActivitePrincipale:
-      libelleFromCodeNAFWithoutNomenclature(activite_principale),
-    activitePrincipale: activite_principale,
-    dateCreation: parseDateCreationInsee(date_creation),
-    dateDebutActivite: date_debut_activite ?? '',
-    dateFermeture: date_fermeture ?? '',
-    complements: {
-      estEntrepreneurIndividuel,
-      idFiness: liste_finess || [],
-      idBio: liste_id_bio || [],
-      idcc: liste_idcc || [],
-      idOrganismeFormation: liste_id_organisme_formation || [],
-      idRge: liste_rge || [],
-      idUai: liste_uai || [],
-    },
   };
 };
