+++ models/rne/dirigeants.tsx
@@ -1,15 +1,16 @@
 import { fetchRNEImmatriculation } from '#clients/api-proxy/rne';
 import { HttpNotFound } from '#clients/exceptions';
+import { clientDirigeantsRechercheEntreprise } from '#clients/recherche-entreprise/dirigeants';
 import { EAdministration } from '#models/administrations/EAdministration';
 import {
   APINotRespondingFactory,
   IAPINotRespondingError,
 } from '#models/api-not-responding';
 import { verifySiren } from '#utils/helpers';
-import { IDirigeants } from '.';
+import { IDirigeants } from './types';
 
 /*
- * Request Immatriculation from INPI's RNE
+ * Request dirigeants from INPI's RNE
  * @param siren
  */
 export const getDirigeantsRNE = async (
@@ -19,11 +20,19 @@ export const getDirigeantsRNE = async (
 
   try {
     const { dirigeants } = await fetchRNEImmatriculation(siren);
-
     return { data: dirigeants, metadata: { isFallback: false } };
   } catch (e: any) {
     if (e instanceof HttpNotFound) {
       return APINotRespondingFactory(EAdministration.INPI, 404);
+    }
+
+    try {
+      const dirigeants = await clientDirigeantsRechercheEntreprise(siren);
+      return { data: dirigeants, metadata: { isFallback: true } };
+    } catch {
+      if (e instanceof HttpNotFound) {
+        return APINotRespondingFactory(EAdministration.INPI, 404);
+      }
     }
 
     return APINotRespondingFactory(EAdministration.INPI, 500);
